Minor changes to the random rama sampling and number of closure attemts in KIC.
This commit reduces the number of closure attempts in each call to the KinematicMover to 500, previous default was 2000. This is considerably faster and only slightly decreases the fraction of sub-Angstrom decoys, RMSDs are virtually unchanged. The number of closure attempts can be controlled with the flag -max_kic_perturber_samples
In addition, the sampling of random phi/psi pairs from rama and rama2b was off by 5deg in the previous implementation, fixed here. Thanks to Kale for spotting this!

All kinematic-closure related integration tests change.

----
An additional 3 commit messages were squashed into this commit:
--
same fix for Rama2b
--
changing the default for -max_kic_perturber_samples from 2000 to 500 - runs faster and gives the same performance in terms of RMSD, and only a small decrease in the percentage of sub-Angstrom decoys with NGK, almost unnoticeable with KIC
--
phi/psi in rama bins are the lower left corner, not the center. Adjusting this in the random rama selection code for KIC.
--

Old repository SHA1: 7999f87f55457eeaf246fa8cf540e494ef5bd949