floating base and variable geometry bug fixes
Fixed a bug in SWM in which virtual riboses were placed on floating bases present in the native pose, leading to crashes in SWM benchmark runs. Code now checks to see if the residue is fixed or not in the domain map.
Additionally, altered variable geometry minimization code in SWA/SWM to only operate on backbone atoms, which fixes cases with distorted base geometries. An alternative fix for the future is to put in improper angle constraints on the planar atoms.

	modified:   source/src/protocols/simple_moves/ConstrainToIdealMover.cc
	modified:   source/src/protocols/swa/StepWiseUtil.cc

Old repository SHA1: 71a4a56cb0b824d738cc9f3a316505380db949bd