Changing the fix for the "abandoned NV" bug.
Basically, cart_bonded and pro_close both have different ways
of scoring the closure of the proline ring.  pro_close uses
a virtual atom, NV, which it tries to keep as close to the
backbone N as possible, and cart_bonded has an energy term
for all bond lenghts and angles and not just the ones along
the atom tree.  Well, if you run a protocol that uses
cartesian minimization for one part (using the cart_bonded
term) and then switch off cart_bonded and turn on pro_close,
you see a huge spike in the energies.  Why?  Because NV
doesn't get assigned a derivative so during cartesian minimization,
NV just stays put.  This means that the rest of the structure
will walk away from NV, so that the distance between N and NV
might end up being several angstroms by the time cartesian
minimization completes.

The original solution was to have the cart_bonded term
update NV's coordinate during setup_for_scoring.  This has
the unfortunate side effect of triggering "No structure mods
during scoring" error messages (incidentally delivered quite
a bit after the structure modification was made).

The new solution is to put a weak force between N and NV,
and now between any "virtual shadow" and the atom it's
shadowing.

5 integration test trajectory changes as a result to the introduction
of this new force into the cart_bonded term; ideally, only NV would
be dragged around, but the force is symmetric, so NV pulls a bit on N.

----
An additional 2 commit messages were squashed into this commit:
--
Adding a weak constraint into the cart_bonded term to pull the NV atom
along with the N atom so that the pro_close term, when evaluated on a
structure following cartesian minimization, will not be sky high.
--
Adding a new piece of data to the residue parameter file: a
virtual shadow, which declares the first atom to be a shadow
of the second.  This will now be interpretted by the cart-bonded
term to mean that a small force should be applied to keep the
shadow atom near the original atom.  Hopefully, this will allow
us to fix the cart_bonded / pro_closure discrepancy bug in a
more graceful way than we are fixing it currently.
--

Old repository SHA1: 1f4528fe2f17a22eee2781cb1167c52924be8550