ProteinResidueConformationFeaturesReporter now always outputs backbone torsions
This was a little bit complicated, as the prior behavior was to output these values as 0.0 if
the output structure was non-idealized. These 0.0 values were used, if the pose was later loaded
from the database, to indicate that backbone torsion angles should not be loaded into the pose,
as it is non-ideal and all, and the cartesian coordinated and side chain chi angles should be used
instead.

This behavior was less than "ideal" (ha ha) as torsion angles are useful in post-run analysis and should
always be output by ProteinResidueConformation. To enable this, I added a new boolean field to
represent if a saved pose is idealized into pose_conformations. This table seems appropriate as it already
contains another boolean column, and comes after a struct_id has already been assigned.

Now, backbone torsions are always output, but only loaded into a pose if the pose_conformations table
indicates that the relevant structure is idealized.

----------------------------------------
Expected unit test status: Pass
Expected integration test changes: 5
These 4 tests change either because of the new pose_conformations schema or because torsions are now output:
    features
    database_jd2_compact_io
    database_jd2_io
    features_database_schema

And of course:
    app_exception_handling

----
An additional 2 commit messages were squashed into this commit:
--
Added new ideal column to pose_conformations table in ab_db_for_tests

This is needed so that the antibody_designer integration test does not fail
--
ProteinResidueConformationFeaturesReporter now always outputs backbone torsions

This was a little bit complicated, as the prior behavior was to output these values as 0.0 if
the output structure was non-idealized. These 0.0 values were used, if the pose was later loaded
from the database, to indicate that backbone torsion angles should not be loaded into the pose,
as it is non-ideal and all, and the cartesian coordinated and side chain chi angles should be used
instead.

This behavior was less than "ideal" (ha ha) as torsion angles are useful in post-run analysis and should
always be output by ProteinResidueConformation. To enable this, I added a new boolean field to
represent if a saved pose is idealized into pose_conformations. This table seems appropriate as it already
contains another boolean column, and comes after a struct_id has already been assigned.

Now, backbone torsions are always output, but only loaded into a pose if the pose_conformations table
indicates that the relevant structure is idealized.

Cherry-picked from kylebarlow/designmixer 6dd6d9c
--

Old repository SHA1: 0abde7f418dc4b65ed72b6a939443cfd9fa6dbf8