Fix the performance of the refactored KIC code.
The performance of the refactored KIC code never matched the performance
of the original KIC code.  This was due to an error in the way the fold
tree was being configured.  The problem is now fixed, but the
architecture is no longer as clean as I would like it to be.  I hope to
push another commit which cleans up the architecture in light of the
fold tree bug (and a few other things) in the next day or so.

----
An additional 29 commit messages were squashed into this commit:
--
Merge remote-tracking branch 'origin/master' into devel

Conflicts:
	source/cmake/build_unit/CMakeLists.txt
--
Tweak my pilot apps and analysis scripts.
--
Setup chainbreak scoring with the KIC fold tree.
--
Add PdbLogger to LoopRelaxMover

It's currently commented out, but this will just make it easier to
quickly turn on in the future.
--
Add KicMover.setup()

This method is used to make sure the fold tree is setup at the right
stage in the loop modeling process.
--
Return the low energy structure for loop modeling.

This is as opposed to just returning the endpoint of the simulation,
which might be much higher in energy than the lowest energy structure
found.
--
Comment on why a custom fold tree is useful.
--
Switch back to using a custom fold tree with KIC.
--
Merge with origin/kalekundert/devel.
--
Tweak the pdb preprocessing script.
--
Add fold-tree unit tests for KIC.
--
Merge with /origin/kalekundert/devel
--
Make numpy_cache use a large blob type.

I also added a check for numpy arrays that are too big to cache in
`helpers.py', so that the scripts will crash gracefully if I ever use
data that's too big in the future.
--
Update the options system for refine_kic_refactor.
--
Use refine_kic_refactor instead of refine_kic_v2.

Just a quick bug fix.
--
Merge with origin/cmake/one_script
--
Rearrange when the cmake modules are included.
--
Prevent useful cmake files from being ignored.
--
Merge with origin/cmake/one-script
--
Add some files I forgot about.
--
Merge with origin/kalekundert/devel
--
Make native_ensemble compatible with MySQL.
--
Add a branch option to make_new_build.

This makes it possible to create builds that are specific to the current
branch, which makes it easier to work with several branches at once.
--
Create a unified cmake build script.
--
Tweak native_ensemble analysis scripts for MySQL.
--
Forgot a header.
--
Update native_ensemble to support MPI.

Some of the analysis scripts also had to be updated to support some of
the additional information I'm now outputting (i.e. temperature).
--
Fix lingering bugs in canonical_sampling.
--
Rename ParallelTempering to MpiParallelTempering.

Did the same thing for the HamiltonianExchange class.  This is a prelude
to making a serial version of the ParallelTempering algorithm.
--

Old repository SHA1: a7863107a182b9f7c402d1b1180734c9290783f0