Add "option file tab completion"
Tab completion is a function of the shell, not of Rosetta, but the form
"@../path/to/optionfile.flags" inhibits tab completion of the filename.

Add "@ ../path/to/optionfile.flags" as a valid alternative. The space
between the @ symbol and the filename allows shells (at least bash and
tcsh) to use tab completion on the filename.

Test score_jd2 should change in commandfile, as I've
changed it to use the new system. The only changes should be the
addition of a space after the @ symbol.

----
An additional 1 commit messages were squashed into this commit:
--
Add "option file tab completion"

Tab completion is a function of the shell, not of Rosetta, but the form
"@../path/to/optionfile.flags" inhibits tab completion of the filename.

Add "@ ../path/to/optionfile.flags" as a valid alternative. The space
between the @ symbol and the filename allows shells (at least bash and
tcsh) to use tab completion on the filename.

Test score_jd2 should change in commandfile and log file, as I've
changed it to use the new system. The only changes should be the
addition of a space after the @ symbol.
--

Old repository SHA1: 910011da7ac22d22ae9fc5e45fef09b81616c9e0