02/27/14 Changing Enzdes to be more campitable with talaris; default scorefxn is now talaris2013_cst, which results in ~2% increase in seq recovery for enzdes benchmarks.
NEWS: In EnzdesBaseProtocol.cc, the default scorefxn is now talaris2013_cst.  To be compatible with talaris2013, the exclude_protein_protein_fa_elec_ bool option in the constructor is now set to false by default (it used to be true, which was necessary when using the enzdes scorefxn so not to double count fa_elec of interface sidechains).  Using talaris with talaris2013_cst.wts gives ~2% increase in sequence recovery on the enzdes benchmark.

 ----------------------------------------
 Unit test status: Failed when I ran it by not because of my changes ("unit_tests             - Broken by Kale Kundert:56400.")
 Integration test status: 3 changed (as expected)

 enzdes: outputs different results/scores due to using new scorefxn and weights (talaris2013_cst)

 inverse_rotamer_remodel: changed the flags file to use talaris_2013_cst.wts; outputs different results/score due to different scorefxn

 ligand_motif_design: I did not touch this; this test uses a custom weights file that is not in the database, so I left it as is. Let me know if this is a problem for anyone (sboyken@gmail.com).

Old repository SHA1: ad2f0791f0f6b8dcafba29718b8a373efebbdbdc