Fix -jd2:ntrials behavior
According to all the documentation, the -ntrials value should be reset
for each nstruct. That is, if you FAIL_RETRY on every attempt, a setting
of -nstruct 6 -jd2:ntrials 5 should run through 30 attempts for each
input structure.

Previously, the ntrials was shared between all nstructs for all input
structures, so a -nstruct 6 -jd2:ntrials 5 with a constant FAIL_RETRY
would do just 10 attempts, 5 for *_0001, and 1 each for *_0002 through
*_0006.

This commit changes it to the intended behavior.

Old repository SHA1: 7eb46d2bf1dfad44045133865235e2da60428018