Fix GasteigerAtomTyper unit tests.
The source of the problem was a *reference* to a COP that was being
passed around. While the underlying object was still existing, the COP
object itself was getting deleted, resulting in bad memory access. This
expressed itself (intermittantly) on Mac but not Linux due to differences in how the
memory was being handled, and if the referenced memory was reused or
not.

Old repository SHA1: f96ec53aff2a575fbe2a4eaf1bad81d4cabee9a0