Fix performance benchmark (ContextDependentGeometricSolEnergy)
ContextDependentGeometricSolEnergy was using the HBondSet but not generating
it. This results in issues when run in a Scorefunction with only those
energy terms and no HBonding terms. Add cautious generation of the HBondSet in
the setup function (along the lines of the HBond terms).

No test changes expected.

Old repository SHA1: 53df8519dbc8cd35e1ef37e72254d9b909e4324a