Fixing a bug that can cause full atom only Docking to ignore constraints. Two integration tests change as a result of caching the constraints to the pose before converting it to the full atom representation. These tests are:
1) docking_distance_constraints
2) docking_site_constraints

The scores and final coordinates of the decoys that are produced by
these tests remain unchanged.

I observed changes in two matcher tests that resulted from swapping HIS
with HIS_D.  This is a weird things that seems to occur sporadically (at
least on my system), and it is unlikely that these tests will change on
the test server.  I wonder if there is anything we can do about that
because it's not great to ignore changing integration tests.

Big thanks to BK for discovering and reporting this bug.

Old repository SHA1: 89295229bb75c443b97a0fe8fa6af406d3e01080