Merge pull request #65 from RosettaCommons/roccomoretti/option_cleanup
Option System Cleanup

Reorganize options_rosetta.py to be a bit more ... organized. It's now ordered: major option groups (in, out, run, jd2, score, packing) first, then other general option groups in alphabetical order, then application-specific option groups in alphabetical order, and then personal/pilot option groups in alphabetical order. (Classification was done by best guess, and is not an official pronouncement. Feel free to reorganize. Also, alphabetical ordering is not absolute - logical regroupings can be permitted.) Some formatting cleanup (tabs, not spaces, despite being a .py file) was also done.

Additionally, approximately 200 options which were never referenced on the C++ level were commented out. Use of these options will now result in an error when you try to run a protocol with them. (Thanks to Luki for help with the clang AST parser, which was used to ferret out unreferenced options.)

Finally, a new option "-run:show_unused_options" was added. When set, at the end of the run a list of options which were user set but never actually referenced by the program will be printed out. The flag is unneeded for JD2 protocols, which will automatically print out the list of user-set-but-unused options at the end of a successful run with at least one output structure. (Thanks to Doug for suggesting this feature.)

No unit test changes, but a large number of integration test changes. These are due either to removed options being removed from the flag file for integration tests, or from the new default-on-in-jd2 behavior of the show_unused_options feature, which touched a large number of tests. (A fair number of non-JD2 integration tests also have user-set-but-unused-options, but are silent in this change.)

Old repository SHA1: 53dc2349b1ede38abdef7b8684e6704ee90e9260