Merge pull request #69 from RosettaCommons/weitzner/stepwise_ccd_slowdown
Weitzner/stepwise ccd slowdown: The CCD update resulted in a slowdown in some cases. Rhiju noticed that the Ramachadran data files were being read in from the disk multiple times, which, in the way it was written before, would happen when constructing a new CCDLoopClosureMover. This problem would then be exacerbated when CCDLoopClosureMover was created on the stack in a for loop.

This revision fixes that by getting a COP to the Ramachandran potential stored in the ScoringManager. Additional changes are introduced to the Ramachandran2B potential to allow arbitrary conformations to be evaluated, and the RamaCheck2B class  is updated to use this new interface. This means you can now use two-body Ramachandran checks during CCD closure!

Finally, this change set updates StepWiseProteinCCD_Closer to have a CCDLoopClosureMoverOP as a data member instead of creating it on the stack to reduce the number of times "new" is called in that protocol.

This should improve the runtime of integration tests that use CCDLoopClosureMover, and all of those tests will change due to the lack of tracer output associated with reading the Ramachandran data files. See http://benchmark.graylab.jhu.edu/revision?id=115&branch= for the full list of changed tests.

Old repository SHA1: 7fcdfd9b796ceb0b86f1cb1a3a7d96279b91dd85