Reverting everything that has happened since 698e461ad265a541.  At this point, the master branch should be exactly where it was prior to the merge of the JK-lab branch with master.  The following should now happen: --bazzoli should fix the JK-lab branch in a branch, and test it by submitting SHAs to the test server (rather than testing it by merging with master over and over) --I'll double-check that the broken Windows PyRosetta build was not due to my changes in vmullig/cycpep, and if my branch is good to go, I'll reinstate its merge with master.  (Note that I had temporarily reinstated the merge of my vmullig/cycpep branch in this branch, but have now rolled that back, so no changes have occurred since 698e461ad265a541).
----
An additional 3 commit messages were squashed into this commit:
--
Reverting more of bazzoli's merges.
--
Reverting the merge of master and JK-lab branch.  It is clear that fixing the compilation errors that this merge created will be a long, laborious, trial-and-error process.  This should be done in a branch, so that the rest of us have a clean version of master against which to test our changes (particularly since OUR merges are now being reverted).  Once the problems have been resolved in your branch, this commit may be reverted and then your branch may be re-merged.
--
Reverting reversion of merge of vmullig/cycpep.  (i.e. RESTORING the changes to GeneralizedKIC, which were not the root of the current problems)
--

Old repository SHA1: c6251f499eb88d67a3f4976ee8c09851599fe7fc