First pass implementation of a MultiThreadedJobDistributor.
For now, the only application that is even remotely safe to use with
the MTJD is rosetta_scripts, and even here the incredible prevalence
of global data that people seem unable to ween themselves from makes
many many rosetta_scripts not safe for use with the MTJD.

To activate the MTJD, use extras=cxx11thread and a compiler that
supports C++11.

Random number generation is now threadsafe.  RandomGenerators should
no longer be constructed in individual .cc files, but instead should
use the single, global RandomGenerator accessible through the function
numeric::random::rg(). I'll email out the python script I used to
replace RandomGenerators so that it's easier for people to fix their
code following this change.

Tracers are now threadsafe. Tracers must now be declared "thread_local"
in addition to "static".  There's currently nothing enforcing this,
unfortunately, so you'll just have to be super mindful for now.

Modest refactoring of the JobDistributor's go_main() function, which was
a mess and is now just a little bit less of a mess.

Changing the access pattern for the JobOutputterObservers so that they
now attach themselves to the JobDistribotor::current_job() instead
of to the JobOutputter (which is not threadsafe).  It looks like
the PoseEvaluators will also need to be revamped in a similar fashion.

----
An additional 38 commit messages were squashed into this commit:
--
Fixing the JobOutputterObserver unit tests following the change in which
the OutputObservers attach themselves to the Job instead of to the
JobOutputter.
--
Merging master into aleaverfay/multithreaded_jd2.  Adding "thread_local"
declarations to several new Tracers (and to some Tracers I missed the first
time 'round).
--
Reworking the JobOutputterObserver system to be threadsafe.  Now the JobOutputterObservers
actually watch the Job -- the JobDistributor::current_job() function already returns
the job that's being run by the current thread -- so that jobs running in separate threads
won't step on each others toes.

Note: the JobOutputters themselves are not threadsafe. If two jobs simultaneously try
mid protocol to write a Pose at the same time, then they can mess each other up.
This commit fixes only the issue with the JobOutputterObservers.
--
Merging in Luki's lock to the JobOutputters into my threadsafe singleton creation changes.
--
Implementing genuinely threadsafe singleton creation for the cxx11thread build
taken from this website:

http://preshing.com/20130930/double-checked-locking-is-fixed-in-cpp11/
--
Adding thread locking in JD2 output observers.
--
Merging master into branch aleaverfay/multithreaded_jd2 and once again
fixing the merge conflicts.
--
Removing one more modification to the DARC code; the code now compiles.
--
Merging master into branch aleaverfay/multithreaded_jd2 following reversions
over the weekend to the DARC code.
--
Adding clang extras=cxx11thread build settings.
--
First pass implementation of a MultiThreadedJobDistributor.

Random number generation is now threadsafe.  RandomGenerators should
no longer be constructed in individual .cc files, but instead should
use the single, global RandomGenerator accessible through the function
numeric::random::rg().

Tracers are now threadsafe. Tracers must now be declared "thread_local"
in addition to "static".

Modest refactoring of the JobDistributor's go_main() function, which was
a mess and is now just a little bit less of a mess.
--
Fixing one last discrepancy in the JobDistributor refactor;
re-enabling a line I commented out in FragmentPicker.cc;
cleaning up a handful of other small things.
--
Fixing the behavior of the JobDistributor when running roseta_scripts
so that the Mover created by the RosettaScriptsParser will be saved for
successive nstruct iterations.
--
Merging master into branch aleaverfay/multithreaded_jd2 ahead of the
pull request to merge this into trunk.  I haven't verified that it
compiles yet.
--
Updating the unit tests following the change to the RandomGenerators.
--
Updating the unit tests for the RandomGenerator and Tracer changes;
more changes to come
--
Fixing the terrible things I had done to the FragmentPicker in order
to get the cxx11thread build working.
--
Making all tracers declared at namespace or global scope thread_local,
which for non-CXX11 non-MULTI_THREAED builds is #defined to nothing.
--
Crippling the behavior of the FragPicker because it doesn't compile
with the cxx11thread build... grrrr

This commit absolutely must be reverted in the future.
--
Prepping for a big replacement of tracers from static to static thread_local.
--
Reverting the TRACER macro modification; pursuing
the thread_local option.
--
Adding a TRACER macro that defines a function containing a
thread_local or static TracerOP so that we can have a bunch
of "global" Tracers that are thread safe.
--
Eliminating a piece of global data from the shape complementarity code
that was producing null pointer errors in multithreaded code.
--
Commenting out the tracer statements to tt.Debug in RotamerSet_.cc
which seem to produce segfaults and maybe even double free errors and
I can't figure out why!  I get consistent segfaults from these tracer
output lines and not from any others.
--
More fixes to make the merge from master work with the RandomGenerator changes.
--
Merging master into branch aleaverfay/multithreaded_jd; bringing over
Doug's fixes to the NormalMode code which has killed my build, it would seem
--
Reworking the tracers so that their visibility isn't computed
just-in-time but instead is computed at the later of these
two points in time:

a) tracer construction
b) core::init

I had been seeing a very strange segfault from outputting
messages to a Tracer, so this was step 1 in making the tracers
a little less thread unsafe.  They are not threadsafe now.
--
Replacing all statically declared RandomGenerator objects with
a single RandomGenerator that's accessible through the function
numeric::random::rg().  The functions numeric::random::uniform()
and numeric::random::gaussian() access the single RandomGenerator
directly.

The RandomGenerator is declared to be "thread_local" when using
the extras=cxx11thread build.  This means that each thread uses
its own random generator.  This avoids the need to lock the
RandomGenerator without leaving open the great problem of having
two threads modifying the internal state of the RandomGenerator
at the same time -- a very real problem that leads to uniform()
returning numbers outside of the range between 0 and 1.
--
Fixing unintended change to devel.test.settings
--
Preliminary commit where I replace all non-static RGs with
static RGs so that they're easier to create regexes for and
find and replace
--
Preliminary reworking of the RandomGenerator class
before editing hundreds of files that declare and use their
own RandomGenerators
--
Modifications to the interface for the Parser/RosettaScriptsParser to
force the creation of a Mover for each job.
--
Progress.  The code now compiles, and links, but it segfaults.
--
Ok, compiling version of the MTJD.
--
First pass implementation of a MultiThreadedJobDistributor.

There are many bells and whistles that it lacks, and this surely
does not compile, and there aren't really any tests for it yet
(though there's a test suite for it where I've been testing things
like the boost::thread_specific_ptr (which requires the boost
libraries to work) and C++11's thread_local storage class specifier).

The framework is in place, though.
--
Fixing the C++11 #inclusion for singleton classes to fix the extras=cxx11thread build.
--
Merging master into aleaverfay/multithreaded_jd2
--
Major refactoring of JobDistributor::go_main to break up this 400 line function
which is quite challenging to read and understand.
--

Old repository SHA1: af243571354122c3460dc42653f76a6df6231f82

Co-authored-by: Luki Goldschmidt <lugo@uw.edu>