Fix instance of shared_from_this from stack object
LoopRelaxMover was allocating a sub-mover on the stack.
The problem is that in the apply of the submover, shared_from_this() was
being called, resulting in an error under our new pointer scheme.

Allocation was changed to be on the heap rather than on the stack.

Thanks to @schaarj for pointing out this issue.

Old repository SHA1: a71179feac2030667e031f5932b500b4ad507a46