Merge pull request #152 from RosettaCommons/Seqeunce_constraints
Applied changes to FavorSequenceProfile mover and SeqprofConsensusOperation taskoperation. The changes allow to the mover and task operation to add seqeunce constraints to chain no. 2 of the pose. The clang test did no run but this is not caused by the changes

Old repository SHA1: 018330193ee5554dba55de35251e7079ff35c9bf