Merge pull request #209 from RosettaCommons/vmullig/fix_silent_read
Adding unit tests demonstrating current major bug with silent files

I'm not sure how to fix this, since I don't understand the recent changes that someone added that created the problem in the first place. Here are the symptoms, though:

- If you import a PDB file, run a Rosetta protocol that changes its size (adding residues to the pose), write the pose out to a silent file, and then try to read the silent file, the read fails with a cryptic error message.
- Binary silent files no longer have REMARK lines for every structure in the file. Instead, there seems to just be one REMARK line at the top of the file. I suspect that this means that information is being lost in multi-pose silent files.
- Related to both, if you import multiple PDB files, add residues to all but the first pose, write out the poses to a silent file (ensuring that the first one written is the one that has NOT had any residues added), then try to read them in, the read is fine.  Failure only occurs if the first entry in a silent file has had residues added.  The problem seems to lie with the FULL_MODEL_PARAMETERS part of the REMARK line, which is repeated only once per silent file and stores information only relating to the first model in the silent file.  (I doubt that this is what whoever added this intended, so he/she should revisit this.  I can't, because the creator of this commented nothing, so I have no idea what functionality I'm supposed to be preserving while fixing this bug.)

I'm therefore adding two unit tests. The first is for the first case (import a PDB, modify the pose length, write a silent file, and read it in), and the second is testing write and read of multi-pose silent files (where the poses are very different). These tests currently fail. This pull request is NOT introducing any bugs (or any changes) to the main code; it is ONLY illustrating a current bug not previously covered by tests.

Old repository SHA1: 8c9e940fcd25c81865c110704e042bdd0a19f5b4