Merge pull request #277 from RosettaCommons/roccomoretti/rotlib_cmp
Add Dunbrack binary validation tests

We've had issues with changes to the codebase which alter the contents of 
the Dunbrack-library-as-read-from-ASCII versus the Dunbrack-library-as-read-
from-a-previous-version's-binary. There's a version number in 
core/pack/dunbrack/RotamerLibrary.cc which is supposed to control for this,
but people don't always realize that their changes require a version number bump.

This commit adds unit and integration tests which attempt to compare the ASCII 
and binary versions of the Dunbrack library, and see if they're different. Note
that this integration test is looking at a single version, rather than a version
comparison.

The test is also not foolproof, as there is a delta in the real number 
comparisons, but hopefully it should catch the major changes.

New unit tests and integration tests - they shouldn't (cross fingers) be failing, 
due to the previous commit which bumped the version number and forced
binary regeneration.

Old repository SHA1: 927458584a05f23fc0a37dd47bc66fc0f3ee7b5b