Merge pull request #316 from RosettaCommons/rpache/force_field_ramping_for_KIC
Force field ramping for refactored KIC and FragmentKIC, and bugfix for the latter.

Adding the flags -loops:ramp_rama and -loops:ramp_fa_rep to the refactored KIC code and fixing a bug in the FragmentKIC code, which had caused a BondAnglePerturber to be used accidentally (a relic from the KIC code refactoring). 
This commit also enables score term ramping for FragmentKIC.

Due to the bugfix, the KIC with fragments integration test is supposed to change.

Old repository SHA1: 8235bdb5731e9dc8471a25bfee9a76a1877f7214