Merge pull request #326 from RosettaCommons/vmullig/remodel_bug
Fixes a bug revealed when RemodelMover is used to introduce disulfide bonds.

The bug is actually in core/conformation/util.cc, in the form_disulfide() function. This function replaces residues with disulfide-linked cysteines (or D-cysteines, or D- or L-homocysteines) and bonds them, but was previously failing to preserve applied variant types. This meant that the cutpoint variants at termini were missing, leading to a crash. In the process of fixing this, I cleaned up some duplicated code, too. I've also added an integration test for placing disulfides with the RemodelMover -- it currently seems to be un-covered by tests.

Old repository SHA1: 1adffffd174f90e9f569f464af74500c197ac7dc