Merge pull request #364 from RosettaCommons/vmullig/valgrind_issue
Vmullig/valgrind issue

I think I've fixed the major issue causing many of my unit and integration tests to fail the Valgrind tests on the test server. It was an issue with order of initialization of variables in the Conformation assignment operator and copy constructor (I think). In the middle of the copy constructor, I was calling a function to which I was passing the new, partially-initialized Conformation object, which was problematic. This has been cleaned up. Testing now.

(This does fix all of the unit and integration tests of mine that were being flagged by Valgrind. No unit or integration test changes.)

Old repository SHA1: afbe193531b71e6ba1c6c2f45e4fc429ef529c0b