Adding a user control for detecting disulfide bonds in SymmetricConformation::detect_disulfides().
The previous behavior triggers the detect_disulfides() no matter what.
With the control now, when -detect_disulf is specified and is false,
this function will not be triggered. The default behavior stays the same -
when the -detect_disulf is not specified, "detect_disulf()" will be
triggered.

Old repository SHA1: f7d1b1882b03ee7b7b8648ae43eb941849dc4945